Add link to solidity install instructions
At this point in the tutorial it is likely that the reader does not have solidity installed, so it probably makes sense to link to where/how to install it. I've included one link I found but if there is a better link that could be included as well. It seems other parts of this page may be out of date but this sentence should help.