Add error handling to Greeter tutorial
While trying to get up and running with contract deployment through geth, I spent multiple hours trying to debug the `greeterFactory.new(...)` command in this tutorial without realizing that I was getting an "Error: exceeds block gas limit" error. This change makes it so that the console will log if contract deployment fails.